projects/gift-registry/requirements.txt
# Core
Django==5.1.*
psycopg2-binary==2.9.*
django-environ==0.11.*

# Celery / Redis
celery==5.4.*
redis==5.2.*
django-celery-results==2.5.*

# Email
django-anymail[mailgun]==12.*

# Storage (S3-compatible: AWS S3 or Cloudflare R2)
django-storages==1.14.*
boto3==1.35.*

# Images
Pillow==11.*

# Forms & UI
django-crispy-forms==2.3.*
crispy-tailwind==1.0.*

# Auth tokens (email verification, password reset, access tokens)
django-allauth==65.*

# Security
django-ratelimit==4.1.*

# Dev tools
django-debug-toolbar==4.4.*
